% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelog.R
\name{baseline_src}
\alias{baseline_src}
\title{Baseline component tables (v1, split without changes)}
\usage{
baseline_src()
}
\value{
A list of data.tables.
}
\description{
Baseline component tables (v1, split without changes)
}
