% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{cc_tables}
\alias{cc_tables}
\alias{cc_forms}
\alias{cc_parts}
\alias{cc_variables}
\alias{cc_xpaths}
\alias{cc_families}
\title{Component tables}
\usage{
cc_forms()

cc_parts()

cc_tables()

cc_variables()

cc_xpaths()

cc_families()
}
\value{
A data.table.
}
\description{
Component tables
}
