% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance990-package.R
\docType{package}
\name{concordance990-package}
\alias{concordance990}
\alias{concordance990-package}
\title{concordance990: IRS 990 E-File Master Concordance}
\description{
The Master Concordance maps xpaths in IRS Form 990 e-file XML returns onto a documented set of research variables and relational tables. This package holds the concordance as component tables (forms, parts, tables, variables, xpaths), builds the unified tidy concordance from them, and provides tools to validate it against filing evidence and to document every change from version 1.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Nonprofit-Open-Data-Collective/concordance990}
}

}
\author{
\strong{Maintainer}: Jesse Lecy \email{jdlecy@gmail.com}

}
\keyword{internal}
