% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{concordance_version}
\alias{concordance_version}
\title{Version of the concordance}
\usage{
concordance_version()
}
\value{
A list with the package version and an md5 fingerprint of the
component tables.
}
\description{
Version of the concordance
}
