% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelog.R
\name{diff_src}
\alias{diff_src}
\title{Cell-level differences between two sets of component tables}
\usage{
diff_src(old, new)
}
\arguments{
\item{old, new}{Lists of component tables (as from \code{\link[=read_src]{read_src()}} or
\code{\link[=baseline_src]{baseline_src()}}).}
}
\value{
A data.table with \code{level}, \code{key}, \code{field}, \code{old_value},
\code{new_value} and \code{change_type} (\code{edit}, \code{add}, \code{remove}, \code{add_column},
\code{remove_column}), in the order the changes must be applied.
}
\description{
Cell-level differences between two sets of component tables
}
