% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{load_evidence}
\alias{load_evidence}
\title{Load evidence tables for reporting}
\usage{
load_evidence(evidence_dir = "evidence")
}
\arguments{
\item{evidence_dir}{Directory written by \code{\link[=combine_evidence]{combine_evidence()}}.}
}
\value{
A list of data.tables.
}
\description{
Load evidence tables for reporting
}
