% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank-meaning.R
\name{money_blank}
\alias{money_blank}
\title{Money flag and blank meaning of each variable}
\usage{
money_blank(cc)
}
\arguments{
\item{cc}{A v2 concordance (one database, see \code{\link[=concordance]{concordance()}}).}
}
\value{
A data.table keyed by \code{rdb_table} and \code{variable_name} with
\code{money_field} and \code{blank_meaning}.
}
\description{
Money flag and blank meaning of each variable
}
\keyword{internal}
