% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{normalize_xpath}
\alias{normalize_xpath}
\title{Normalize raw xpaths for lookup}
\usage{
normalize_xpath(x)
}
\arguments{
\item{x}{Character vector of xpaths.}
}
\description{
Removes repeat indexes (\verb{[1]}) and the \verb{irs:} / \verb{efile:} namespace
prefixes, and makes the path start with a single \code{/}.
}
