% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_cc_csv}
\alias{read_cc_csv}
\title{Read a concordance CSV exactly as stored}
\usage{
read_cc_csv(path)
}
\arguments{
\item{path}{Path to a CSV file.}
}
\value{
A data.table of character columns.
}
\description{
Read a concordance CSV exactly as stored
}
