% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelog.R
\name{read_changelog}
\alias{read_changelog}
\title{Read the change log}
\usage{
read_changelog(path = changelog_path())
}
\arguments{
\item{path}{Path to \code{changes.csv}.}
}
\description{
Read the change log
}
