% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpath-versions.R
\name{sort_versions}
\alias{sort_versions}
\title{Sort schema version strings (\verb{2010v3.2}) by year, then version number}
\usage{
sort_versions(v)
}
\description{
Sort schema version strings (\verb{2010v3.2}) by year, then version number
}
\keyword{internal}
