% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{table_names}
\alias{table_names}
\title{Table names}
\usage{
table_names(cardinality = NULL)
}
\arguments{
\item{cardinality}{\code{NULL} for all tables, or \code{"ONE"} / \code{"MANY"}.}
}
\value{
Character vector of \code{rdb_table} names.
}
\description{
Table names
}
