% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\docType{data}
\name{v1_columns}
\alias{v1_columns}
\title{Columns of the v1 concordance, in order}
\format{
An object of class \code{character} of length 23.
}
\usage{
v1_columns
}
\description{
Columns of the v1 concordance, in order
}
\keyword{datasets}
