% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{write_src}
\alias{write_src}
\title{Write component tables}
\usage{
write_src(s, out_dir = src_path())
}
\arguments{
\item{s}{A list of component tables.}

\item{out_dir}{Output directory.}
}
\description{
Write component tables
}
