% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{xpath_map}
\alias{xpath_map}
\title{Lookup table from xpath to variable and table}
\usage{
xpath_map()
}
\value{
A data.table with \code{xpath}, \code{variable_name}, \code{rdb_table},
\code{rdb_relationship} and \code{form_type}.
}
\description{
The minimal columns a parser needs.
}
